requirements.txt
BeautifulSoup==3.2.1
Pillow==2.3.0
PyYAML==3.10
argparse==1.2.1
beanstalkc==0.4.0
beautifulsoup4==4.3.2
cssselect==0.9.1
feedparser==5.1.3
goose-extractor==1.0.8
jieba==0.32
lxml==3.3.3
nltk==2.0.4
pymongo==2.6.3
requests==2.2.1
wsgiref==0.1.2
html5lib
fabric
python-dateutil
mixpanel-py
